pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mimeta-pytorch"
version = "0.0.2"
authors = [
    { name="Stefano Woerner", email="stefano@woerner.eu" },
]
description = "Library for using the MIMeta dataset"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "h5py~=3.8.0",
    "numpy~=1.24.3",
    "torch~=2.0.0",
    "PyYAML~=6.0",
    "Pillow~=9.5.0",
    "torchvision~=0.15.1",
    "torchcross~=0.0.2",
]

[project.urls]
"Homepage" = "https://github.com/StefanoWoerner/mimeta-pytorch"
"Issue Tracker" = "https://github.com/StefanoWoerner/mimeta-pytorch/issues"
